% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetheme.R
\name{basetheme}
\alias{basetheme}
\title{Base Theme}
\usage{
basetheme(...)
}
\arguments{
\item{...}{- a sequence of \code{parameter=value} pairs (see Details).}
}
\value{
a list of all theme settings (invisibly, unless no arguments were provided).
}
\description{
Sets themes for base plotting
}
\details{
Function dispatches based on the type of first argument:\cr\cr
1. No arguments - the current theme settings are returned.\cr
2. NULL - all theme settings are removed.\cr
3. list - assumed that list stores theme settings.\cr
4. character - a theme with that name is used.\cr

Further arguments can be provided as \code{parameter=value} pairs.
See examples.
}
\examples{
  basetheme("subtle")
  plot(1)
  basetheme(basetheme_clean())
  pairs(iris[,1:4], col=iris$Species)
  basetheme("clean", rect.col="grey90")
  pairs(iris[,1:4], col=iris$Species)
  basetheme()
  plot(1, 1)
  basetheme(NULL)
  dev.off(); plot(1, 1)

}
\seealso{
\code{basetheme_default()}
}
\author{
Karolis Koncevičius
}
