% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltools.R
\name{colshade}
\alias{colshade}
\title{Add Shade to Color}
\usage{
colshade(cols, frac = 0)
}
\arguments{
\item{cols}{a vector of colors}

\item{frac}{a vector of shade fractions (between -1 and 1, defaults to 0)}
}
\value{
a vector of colors with added shades or tints.
}
\description{
Adds a selected amount of shade or tint to a vector of colors.
}
\details{
This function adds shades and tints to provided list of colors.
Shade or tint is decided depending on the sign of the \code{frac} argument:
positive values make colors darker and negative values lighten them.
}
\author{
Karolis Koncevičius
}
