% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettheme.R
\name{gettheme}
\alias{gettheme}
\title{Get Theme}
\usage{
gettheme(theme = NULL)
}
\value{
a list of theme parameters used by \code{settheme()} along with their settings.
}
\description{
Obtains a list of parameter-value pairs for a selected theme.
}
\details{
The parameters used within each theme consists of:\cr
1) A subset of \code{par} parameters that are most universal to all plots.\cr
2) "palette" - a vector of colors that will be used when colors are specified by numbers.\cr
3) rect.* - \code{rect()} parameters that will be applied to the background of the plotting region.\cr\cr

Calling the \code{gettheme()} function with no arguments will return the settings of the current theme.
Calling the \code{gettheme()} function with a name of selected theme returns a list of all parameter values for that theme.
}
\examples{
  gettheme()
  gettheme("clean")

}
\seealso{
\code{settheme()}
}
\author{
Karolis Koncevičius
}
