% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltools.R
\name{lab2col}
\alias{lab2col}
\title{Labels to Colors}
\usage{
lab2col(x, pal, ref = x, NAcol)
}
\arguments{
\item{x}{vector of labels (always transformed to character)}

\item{pal}{colors used to build the palette (defaults to colors set by theme)}

\item{ref}{reference for assigning colors (defaults to elements of \code{x})}

\item{NAcol}{color to be used for NA values (defaults to color set by theme)}
}
\value{
a vector of colors for each element in x.
}
\description{
Assigns colors to the provided vector of labels.
}
\details{
This function assigns colors to each unique level of label vector \code{x}.
Mainly used for consistently assigning colors to a unique set of labels,
especially when the order of original labels might change (see examples).

Color of NA values and values outside of specified reference can be set using
\code{NAcol} argument. Set this to NA to omit the display of such values.

\code{ref} parameter can be specified to select the order of color assignment:
first element from \code{ref} will be assigned the first color from \code{pal},
second element - second color, and so on.
}
\examples{
 pairs(iris[,1:4], col=lab2col(iris$Species))

 labs1 <- c("A", "B", "C", "D", "E")
 labs2 <- c("D", "C", "E", "A", "B")
 labs3 <- c("A", "A", "C", "F", "B")
 par(mfrow=c(2,2))
 barplot(1:5, names=labs1, col=lab2col(labs1))
 barplot(1:5, names=labs2, col=lab2col(labs2, ref=labs1))                # color orders are preserved
 barplot(1:5, names=labs3, col=lab2col(labs3, ref=labs1))                # new levels are treated as missing
 barplot(1:5, names=labs3, col=lab2col(labs3, ref=labs1, NAcol="black")) # "missing" color can be adjusted

 ref <- sort(unique(mtcars$cyl))
 col <- c("pink", "red", "darkred")
 plot(mtcars$hp, mtcars$mpg, col=lab2col(mtcars$cyl, col, ref), pch=19)
 legend("topright", legend=ref, fill=col, title="cylinders")

}
\author{
Karolis KonceviÄius
}
