% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settheme.R
\name{settheme}
\alias{settheme}
\title{Set Theme}
\usage{
settheme(...)
}
\arguments{
\item{...}{- a sequence of \code{parameter=value} pairs (see Details).}
}
\value{
a list of all theme settings (invisibly, unless no arguments were provided).
}
\description{
Sets themes for base plotting
}
\details{
Function dispatches based on the type of first argument:\cr\cr
1. No arguments - all theme settings are removed.\cr
2. NULL - all theme settings are removed.\cr
3. list - assumed that list stores theme settings.\cr
4. character - a theme with that name is used.\cr

Further arguments can be provided as \code{parameter=value} pairs.
See examples.
}
\examples{
  settheme("subtle")
  plot(1)
  settheme(gettheme("clean"))
  pairs(iris[,1:4], col=iris$Species)
  settheme("clean", rect.col="grey90")
  pairs(iris[,1:4], col=iris$Species)
  plot(1, 1)
  settheme(NULL)
  dev.off(); plot(1, 1)

}
\seealso{
\code{gettheme()}
}
\author{
Karolis Koncevičius
}
