% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2col.R
\name{num2col}
\alias{num2col}
\title{Numeric Vector to Color}
\usage{
num2col(x, ..., xmin = min(x, na.rm = TRUE), xmax = max(x, na.rm =
  TRUE))
}
\arguments{
\item{x}{numeric vector}

\item{...}{colors to build palette from (defaults to jet colors)}

\item{min}{minimum range value (defaults to \code{min(x)})}

\item{max}{maximum range value (defaults to \code{max(x)})}
}
\value{
a vector of colors for each element in x.
}
\description{
Returns interpolated colors for a specified numeric vector.
}
\details{
This function interpolates a given set of colors to a numeric vector.
Main use case is in turning numbers into colors for plots,
especially when different ranges of \code{x} has to be colored differently.
}
\examples{
  plot(mtcars$hp, mtcars$mpg, col=num2col(mtcars$mpg), pch=19, cex=2)

  cols <- c("cornflowerblue", "orange")
  plot(mtcars$hp, mtcars$mpg, col=num2col(mtcars$mpg, cols), pch=19, cex=2)

  plot(mtcars$hp, mtcars$mpg, col=num2col(mtcars$mpg, "green", "red"), pch=19, cex=2)

}
\author{
Karolis KonceviÄius
}
