% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{basetheme_default}
\alias{basetheme_dark}
\alias{basetheme_minimal}
\alias{basetheme_void}
\alias{basetheme_brutal}
\alias{basetheme_clean}
\alias{basetheme_ink}
\alias{basetheme_subtle}
\title{Theme Functions}
\usage{
basetheme_default()

basetheme_dark()

basetheme_minimal()

basetheme_void()

basetheme_brutal()

basetheme_clean()

basetheme_ink()

basetheme_subtle()
}
\value{
a list of parameters used by \code{basetheme} along with their settings.
}
\description{
A list of available theme functions.
}
\details{
For now the following themes are implemented:
1. \code{basetheme_default()}\cr
2. \code{basetheme_clean()}\cr
3. \code{basetheme_brutal()}\cr
4. \code{basetheme_dark()}\cr
4. \code{basetheme_ink()}\cr
5. \code{basetheme_subtle()}\cr
5. \code{basetheme_minimal()}\cr
6. \code{basetheme_void()}\cr

The used parameters are:
1) A subset of \code{par} parameters that are most universal to all plots.
2) "palette" - a vector of colors that will be used when colors are specified by numbers.
3) rect.* - \code{rect()} parameters that will be applied to the background of the plotting region.
}
\examples{
  basetheme_default()
  basetheme_clean()
  basetheme_brutal()
  basetheme_dark()
  basetheme_ink()
  basetheme_subtle()
  basetheme_minimal()
  basetheme_void()

}
\seealso{
\code{basetheme}
}
\author{
Karolis Koncevičius
}
